+++ src/node/behaviors/__init__.py
@@ -29,13 +29,13 @@ from .mapping import FullMapping  # noqa
 from .mapping import ItemMapping  # noqa
 from .mapping import IterableMapping  # noqa
 from .mapping import Mapping  # noqa
+from .mapping import MappingNode  # noqa
+from .mapping import Nodify  # noqa
 from .mapping import ReadMapping  # noqa
 from .mapping import WriteMapping  # noqa
-from .nodespace import Nodespaces  # noqa
 from .node import DefaultInit  # noqa
-from .node import MappingNode  # noqa
 from .node import Node  # noqa
-from .node import Nodify  # noqa
+from .nodespace import Nodespaces  # noqa
 from .order import Order  # noqa
 from .reference import NodeIndex  # noqa
 from .reference import Reference  # noqa